everything completely refactored apart form PropertiesPanel.jsx
