+++ client/src/components/content/PropertiesPanel/PropertiesPanel.css
@@ -11,3 +11,7 @@
   margin-left: 100;
   padding-bottom: 1rem;
 }
+.properties-panel-dropdown {
+  width: 100%;
+  margin-top: 10px;
+}
