--- client/src/components/content/PropertiesPanel/PropertiesPanelBuilder/PropertiesPanelBuilder.jsx
@@ -1,353 +1,0 @@
-import { is } from 'bpmn-js/lib/util/ModelUtil';
-
-import React, { Component } from 'react';
-import PropertiesPanelApplicationDropdown from '../PropertiesPanelApplicationDropdown/PropertiesPanelApplicationDropdown';
-import PropertiesPanelTaskDropdown from '../PropertiesPanelTaskDropdown/PropertiesPanelTaskDropdown';
-
-import { Button, Input, Tooltip, Typography } from 'antd';
-import { InfoCircleOutlined, RobotOutlined } from '@ant-design/icons';
-
-import '../PropertiesView/PropertiesView.css';
-
-const activityDataRetrieval = require('../../../../utils/xmlUtils');
-
-const { Text } = Typography;
-
-/**
- * @class
- * @component
- * @classdesc Builds PropertiesPanel for one selected BPMN-Element.
- * @example return (<PropertiesPanelBuilder />)
- *
- * @description Initializes state based on properties; initializes session storage. Binds all state-methods.
- */
-export default class PropertiesPanelBuilder extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      element: props.element,
-      modeler: props.modeler,
-      selectedApplication: '',
-      tasksForSelectedApplication: [''],
-      disableTaskSelection: true,
-    };
-
-    let { element } = this.state;
-    if (element.labelTarget) {
-      this.setState({
-        element: element.labelTarget,
-      });
-    }
-
-    this.updateSelectedApplication = this.updateSelectedApplication.bind(this);
-    this.updateSelectedTask = this.updateSelectedTask.bind(this);
-    this.makeServiceTask = this.makeServiceTask.bind(this);
-
-    this.initSessionStorage('TaskToApplicationCache', JSON.stringify({}));
-    this.initSessionStorage('AvailableApplications', []);
-  }
-
-  /**
-   * @description Checks if passed item already exists in session storage and initializes with given value if not existing.
-   * @param {*} itemToCheckFor The selected item to check for in the session storage.
-   * @param {*} valueToInitTo The value to init to if the item is not existing in session storage yet.
-   */
-  initSessionStorage(itemToCheckFor, valueToInitTo) {
-    if (sessionStorage.getItem(itemToCheckFor) === null)
-      sessionStorage.setItem(itemToCheckFor, valueToInitTo);
-  }
-
-  /**
-   * @description
-   * Checks if available application list hast just been initialized
-   * or already populated with values and does so if not done yet.
-   */
-  componentDidMount() {
-    let applicationList = sessionStorage.getItem('AvailableApplications');
-    if (applicationList.length < 1)
-      this.saveAvailableApplicationsToSessionStorage();
-
-    this.checkForExistingRPAAttributes();
-  }
-
-  /**
-   * @description gets called each time the Component is mounted to ensure that if an
-   * element already has an app and Task selected, those will be displayed
-   */
-  checkForExistingRPAAttributes() {
-    let { element } = this.state;
-    if (element.businessObject['$attrs']['arkRPA:application'])
-      this.getTasksForApplication(
-        element.businessObject['$attrs']['arkRPA:application']
-      );
-
-    if (
-      !element.businessObject['$attrs']['arkRPA:application'] &&
-      this.state['disableTaskSelection']
-    ) {
-      this.setState({ disableTaskSelection: true });
-    } else {
-      this.setState({ disableTaskSelection: false });
-    }
-  }
-
-  /**
-   * @description Fetch all applications from MongoDB and save in session storage.
-   */
-  async saveAvailableApplicationsToSessionStorage() {
-    await fetch('/rpa-framework/commands/get-available-applications')
-      .then((response) => response.json())
-      .then((data) => {
-        sessionStorage.setItem('AvailableApplications', data);
-      });
-  }
-
-  /**
-   * @description
-   * Checks if tasks for selected application are already stored in session storage.
-   * Otherwise, fetch tasklist from MongoDB.
-   * @param {*} selectedApplication Application for which to get the tasks for.
-   */
-  async getTasksForApplication(selectedApplication) {
-    let currentSavedTasksObject = JSON.parse(
-      sessionStorage.getItem('TaskToApplicationCache')
-    );
-
-    if (selectedApplication in currentSavedTasksObject) {
-      this.setState({
-        selectedApplication: selectedApplication,
-        tasksForSelectedApplication:
-          currentSavedTasksObject[selectedApplication],
-        disableTaskSelection: false,
-      });
-    } else {
-      this.fetchTasksFromDB(selectedApplication, currentSavedTasksObject);
-    }
-  }
-
-  /**
-   * @description Fetch tasklist from Mongo-DB and set state to force rerendering.
-   * @param {String} selectedApplication - String with currently selected application from Dropdown
-   * @param {Object} currentSavedTasksObject - Object with all applications and tasks as attributes
-   */
-  async fetchTasksFromDB(selectedApplication, currentSavedTasksObject) {
-    await fetch(
-      '/rpa-framework/commands/get-available-tasks-for-application?application=' +
-        selectedApplication.replaceAll(' ', '+')
-    )
-      .then((response) => response.json())
-      .then((data) => {
-        currentSavedTasksObject[selectedApplication] = data;
-        sessionStorage.setItem(
-          'TaskToApplicationCache',
-          JSON.stringify(currentSavedTasksObject)
-        );
-
-        this.setState({
-          tasksForSelectedApplication: data,
-          disableTaskSelection: false,
-        });
-      });
-  }
-
-  updateName(name) {
-    const modeling = this.state['modeler'].get('modeling');
-    let { element } = this.state;
-    modeling.updateLabel(element, name);
-  }
-
-  updateTopic(topic) {
-    const modeling = this.state['modeler'].get('modeling');
-
-    let { element } = this.state;
-    modeling.updateProperties(element, {
-      'custom:topic': topic,
-    });
-  }
-
-  updateSelectedApplication(value) {
-    this.setState(
-      {
-        selectedApplication: value,
-      },
-      () => this.getTasksForApplication(value)
-    );
-  }
-
-  updateSelectedTask(value) {
-    this.setState(
-      {
-        selectedTask: value,
-      },
-      () => {
-        const modeling = this.state['modeler'].get('modeling');
-        let { element } = this.state;
-        activityDataRetrieval.fetchAndUpdateRPAProperties(
-          this.state['selectedApplication'],
-          value,
-          modeling,
-          element
-        );
-      }
-    );
-  }
-
-  makeServiceTask(name) {
-    const bpmnReplace = this.state['modeler'].get('bpmnReplace');
-
-    let { element } = this.state;
-    bpmnReplace.replaceElement(element, {
-      type: 'bpmn:ServiceTask',
-    });
-  }
-
-  attachTimeout() {
-    const modeling = this.state['modeler'].get('modeling');
-    const selection = this.state['modeler'].get('selection');
-
-    let { element } = this.state;
-    const attrs = {
-      type: 'bpmn:BoundaryEvent',
-      eventDefinitionType: 'bpmn:TimerEventDefinition',
-    };
-
-    const position = {
-      x: element.x + element.width,
-      y: element.y + element.height,
-    };
-
-    const boundaryEvent = modeling.createShape(attrs, position, element, {
-      attach: true,
-    });
-
-    const taskShape = this.append(boundaryEvent, {
-      type: 'bpmn:Task',
-    });
-
-    selection.select(taskShape);
-  }
-
-  isTimeoutConfigured(element) {
-    const attachers = element.attachers || [];
-
-    return attachers.some((e) => hasDefinition(e, 'bpmn:TimerEventDefinition'));
-  }
-
-  append(element, attrs) {
-    const autoPlace = this.state['modeler'].get('autoPlace');
-    const elementFactory = this.state['modeler'].get('elementFactory');
-
-    var shape = elementFactory.createShape(attrs);
-
-    return autoPlace.append(element, shape);
-  }
-
-  render() {
-    return (
-      <>
-        <div className='element-properties' key={this.props.element.id}>
-          <fieldset>
-            {is(this.props.element, 'bpmn:Task') && (
-              <Text
-                className='label-on-dark-background'
-                style={{ fontSize: '24pt' }}
-              >
-                Activity
-              </Text>
-            )}
-            {is(this.props.element, 'bpmn:Event') && (
-              <Text
-                className='label-on-dark-background'
-                style={{ fontSize: '24pt' }}
-              >
-                Event
-              </Text>
-            )}
-            {is(this.props.element, 'bpmn:Gateway') && (
-              <Text
-                className='label-on-dark-background'
-                style={{ fontSize: '24pt' }}
-              >
-                Gateway
-              </Text>
-            )}
-          </fieldset>
-
-          <fieldset>
-            <Text className='label-on-dark-background'>ID: </Text>
-            <Text className='label-on-dark-background'>
-              {this.props.element.id}
-            </Text>
-          </fieldset>
-
-          <fieldset>
-            <Text className='label-on-dark-background'>Name:</Text>
-            <Input
-              placeholder='name'
-              style={{ marginBottom: '10px' }}
-              /*prefix={<UserOutlined className="site-form-item-icon" />}*/
-              suffix={
-                <Tooltip title='the name of your task, gateway or event'>
-                  <InfoCircleOutlined style={{ color: 'rgba(0,0,0,.45)' }} />
-                </Tooltip>
-              }
-              value={this.props.element.businessObject.name || ''}
-              onChange={(event) => {
-                this.updateName(event.target.value);
-              }}
-            />
-          </fieldset>
-
-          <fieldset>
-            {is(this.props.element, 'bpmn:Task') && (
-              <>
-                <Text className='label-on-dark-background'>Actions: </Text>
-                <br />
-                <Button
-                  type='primary'
-                  onClick={this.makeServiceTask}
-                  icon={<RobotOutlined />}
-                  style={{
-                    width: '80%',
-                    marginTop: '10px',
-                    marginLeft: '30px',
-                    marginRight: '30px',
-                  }}
-                >
-                  Make RPA Task
-                </Button>
-                <PropertiesPanelApplicationDropdown
-                  onApplicationSelection={this.updateSelectedApplication}
-                  applications={sessionStorage
-                    .getItem('AvailableApplications')
-                    .split(',')}
-                  currentSelection={
-                    this.props.element.businessObject['$attrs'][
-                      'arkRPA:application'
-                    ]
-                  }
-                />
-                <br />
-                <PropertiesPanelTaskDropdown
-                  listOfTasks={this.state['tasksForSelectedApplication']}
-                  onTaskSelection={this.updateSelectedTask}
-                  disabled={this.state['disableTaskSelection']}
-                  currentSelection={
-                    this.props.element.businessObject['$attrs']['arkRPA:task']
-                  }
-                />
-              </>
-            )}
-          </fieldset>
-        </div>
-      </>
-    );
-  }
-}
-
-// helpers ~ legacy ///////////////////
-function hasDefinition(event, definitionType) {
-  const definitions = event.businessObject.eventDefinitions || [];
-
-  return definitions.some((d) => is(d, definitionType));
-}
