+++ client/src/components/content/PropertiesPanel/PropertiesPanelView/PropertiesPanelView.module.css
@@ -1,0 +1,18 @@
+.text {
+  font-size: 24px !important;
+}
+
+.input {
+  margin-bottom: 10px;
+}
+
+.infoCircleOutlined {
+  color: rgba(0, 0, 0, 0.45);
+}
+
+.button {
+  width: 80%;
+  margin-top: 10px;
+  margin-left: 30px;
+  margin-right: 30px;
+}
